Move PluginDelayableHook to me.ketal.base.*
